+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to
 
 ### Added
 
-- New `kindle-sync` command for syncing a user's kindle highlights and notes to their vault
+- New `kindle-sync` command for syncing a user's kindle highlights and notes to
+  their vault.
 
 ## [0.13.1] - 2025-10-13